+++ configs/http/middleware.php
@@ -3,6 +3,7 @@
  * Defines the list of global middleware to be run on all routes
  */
 use Project\HTTP\Middleware\Session;
+use Project\HTTP\Middleware\CheckCSRFToken;
 
 /**
  * ----------------------------------------------------------
@@ -17,5 +18,6 @@ return [
      *
      * List any HTTP middleware you'd like here
      */
-    Session::class
+    Session::class,
+    CheckCSRFToken::class
 ];